+++ functions.php
@@ -46,7 +46,7 @@ final class BFG {
 
 		// Never show errors on production
 		// https://stackoverflow.com/questions/9242903/php-hide-all-errors
-		if( BFG_PRODUCTION ) {
+		if( BFG_PRODUCTION && !defined('WP_DEBUG') ) {
 			error_reporting(0);
 			ini_set('display_errors', 0);
 		}