+++ includes/admin/admin-functions.php
@@ -20,59 +20,6 @@ function bfg_load_admin_assets() {
 
 }
 
-add_action( 'pre_ping', 'bfg_disable_self_pings' );
-/**
- * Prevent the child theme from being overwritten by a WordPress.org theme with the same name.
- *
- * See: http://wp-snippets.com/disable-self-trackbacks/
- *
- * @since 2.0.0
- */
-function bfg_disable_self_pings(&$links) {
-
-	foreach ( $links as $l => $link )
-		if ( mb_strpos( $link, home_url() ) === 0 )
-			unset($links[$l]);
-
-}
-
-/**
- * Change WP JPEG compression (WP default is 90%).
- *
- * See: http://wpmu.org/how-to-change-jpeg-compression-in-wordpress/
- *
- * @since 2.0.14
- */
-// add_filter( 'jpeg_quality', 'bfg_set_jpeg_quality' );
-function bfg_set_jpeg_quality() {
-
-	return 80;
-
-}
-
-/**
- * Disable srcset.
- *
- * @since 20200420
- */
-// add_filter( 'wp_calculate_image_srcset', '__return_false' );
-
-/**
- * Disable lazy loading.
- *
- * @since 20200811
- */
-// add_filter( 'wp_lazy_loading_enabled', '__return_false' );
-
-/**
- * Add new image sizes.
- *
- * See: http://wptheming.com/2014/04/features-wordpress-3-9/
- *
- * @since 2.0.0
- */
-// add_image_size( 'desktop-size', 1024, 768, array( 'left', 'top' ) ); // Crop positions are: top, left, right, bottom, center
-
 // add_filter( 'upload_mimes', 'bfg_enable_svg_uploads', 10, 1 );
 /**
  * Enabled SVG uploads. Note that this could be a security issue, see: https://bjornjohansen.no/svg-in-wordpress.
@@ -85,49 +32,6 @@ function bfg_enable_svg_uploads($mimes) {
 	$mimes['svgz'] = 'image/svg+xml';
 
 	return $mimes;
-
-}
-
-// add_filter( 'image_size_names_choose', 'bfg_image_size_names_choose' );
-/**
- * Add new image sizes to media size selection menu.
- *
- * See: http://wpdaily.co/top-10-snippets/
- *
- * @since 2.0.0
- */
-function bfg_image_size_names_choose($sizes) {
-
-	$sizes['desktop-size'] = __( 'Desktop', CHILD_THEME_TEXT_DOMAIN );
-
-	return $sizes;
-
-}
-
-/*
- * Activate the Link Manager
- *
- * See: http://wordpressexperts.net/how-to-activate-link-manager-in-wordpress-3-5/
- *
- * @since 2.0.1
- */
-// add_filter( 'pre_option_link_manager_enabled', '__return_true' );		// Activate
-
-/*
- * Disable pingbacks
- *
- * See: http://wptavern.com/how-to-prevent-wordpress-from-participating-in-pingback-denial-of-service-attacks
- *
- * Still having pingback/trackback issues? This post might help: https://wordpress.org/support/topic/disabling-pingbackstrackbacks-on-pages#post-4046256
- *
- * @since 2.2.3
- */
-add_filter( 'xmlrpc_methods', 'bfg_remove_xmlrpc_pingback_ping' );
-function bfg_remove_xmlrpc_pingback_ping($methods) {
-
-	unset($methods['pingback.ping']);
-
-	return $methods;
 
 }
 
@@ -162,13 +66,6 @@ function bfg_remove_files_on_upgrade() {
  * @since 20170608
  */
 add_filter( 'secure_signon_cookie', '__return_true' );
-
-/*
- * Prevent login with username (email only).
- *
- * @since 20180604
- */
-// remove_filter( 'authenticate', 'wp_authenticate_username_password', 20 );
 
 /*
  * Prevent non-SSL HTTP origins.