+++ includes/genesis.php
@@ -3,13 +3,6 @@
 if( !defined( 'ABSPATH' ) ) exit; // Exit if accessed directly
 
 /*
- * Remove the Genesis redirect on theme upgrade
- *
- * @since 2.3.29
- */
-remove_action( 'genesis_upgrade', 'genesis_upgrade_redirect' );
-
-/*
  * Force HTML5
  *
  * See: http://www.briangardner.com/code/add-html5-markup/
@@ -35,24 +28,6 @@ add_theme_support( 'genesis-accessibility', array('404-page', 'headings', 'scree
 remove_theme_support( 'genesis-import-export-menu' );
 
 /**
- * Add Genesis post format support.
- *
- * @since 2.0.9
- */
-// add_theme_support( 'post-formats', array(
-// 	'aside',
-// 	'chat',
-// 	'gallery',
-// 	'image',
-// 	'link',
-// 	'quote',
-// 	'status',
-// 	'video',
-// 	'audio'
-// ));
-// add_theme_support( 'genesis-post-format-images' );
-
-/**
  * Add support for after entry widget.
  *
  * @since 2.3.33
@@ -65,19 +40,6 @@ remove_theme_support( 'genesis-import-export-menu' );
  * @since 2.0.1
  */
 // add_theme_support( 'genesis-footer-widgets', 3 );
-
-/**
- * Add Genesis theme color scheme selection theme option.
- *
- * @since 2.0.11
- */
-// add_theme_support(
-// 	'genesis-style-selector',
-// 	array(
-// 		'bfg-red' => 'Red',
-// 		'bfg-orange' => 'Orange'
-// 	)
-// );
 
 /**
  * Declare WooCommerce support, using Genesis Connect for WooCommerce.