+++ js/_partials/_ajax.js
@@ -1,9 +1,5 @@
 /* globals ajaxurl */
 
-/* eslint compat/compat:0 */
-
-import URLSearchParams from '@ungap/url-search-params';
-
 /**
  * Sends an AJAX request to WP's admin-ajax.php endpoint
  * @param {Object} config - The configuration details for this request.
@@ -23,18 +19,6 @@ export function ajax(config) {
 		on_error,
 		on_complete
 	} = config;
-
-	if( !(`fetch` in window) ) {
-		if (on_error) {
-			on_error(`Unable to fetch data. Please reload the page and try again.`);
-		}
-
-		if (on_complete) {
-			on_error(`Unable to fetch data. Please reload the page and try again.`);
-		}
-
-		return;
-	}
 
 	const url = typeof ajaxurl === `string` ? ajaxurl : document.body.dataset.ajax_url;