+++ js/scripts.js
@@ -1,5 +1,3 @@
-import './_svgs.js';
-
 // import im from 'include-media-export/include-media.js';
 // import fitvids from 'fitvids/dist/fitvids.js';